requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
redis==5.0.8
grpcio==1.66.1
grpcio-tools==1.66.1
protobuf==5.28.2
